refactor(profile): streamline profile page structure and component imports
- Removed unused imports and interfaces from the profile page.
- Introduced new context and hooks for better state management.
- Updated component imports to reflect the new structure.
- Enhanced tab management with initial tab state based on URL parameters.

This refactor aims to improve code organization and maintainability.
